archive-release: make the new commits consistent
We set the author and committer name and email to a specific value, rather
than that of the build user, and for existing git repositories, we set the
author and committer dates to that of the current HEAD commit. Also drop the
distro version from the commit, so it doesn't change just due to a minor
version bump. This should make the commit ids consistent, and therefore the
pack ids consistent, between different builds for layers which haven't
changed.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>